Fix the PickleError with RawDeltaTable.
